sidechains: one Core client every fork reads
